add "long benchmarks" for KeyValueStore testing
These will only be run when -test-bench-long is added to the test flags.

A large corpus of paths matching a natural ("real-life") hierarchy is
read from paths.data.gz and imported into a particular KeyValueStore.
Recursive and non-recursive queries are run on it to detect performance
problems that arise only at scale.

I generated paths.data.gz by piping the output of `sudo find / -type f`
on my workstation into a python tool I made for the purpose. It changes
the names of all directories and files from my system to random english
words. I'm checking that tool in here as
transform_paths_to_random_words.py, but I plan to remove it from the
repo again once it's in the history and properly findable.